# docs/killrvideo_openapi.yaml
openapi: 3.1.0
info:
  title: KillrVideo v2 - Monolith Backend
  version: 2.0.0-alpha-local
paths:
...
  /api/v1/videos:
    post:
      tags:
      - Videos
      summary: Submit YouTube URL (async processing)
      description: 'Submit a new YouTube video for asynchronous processing.


        Only users with the *creator* or *moderator* role are authorized.'
      operationId: submit_video_api_v1_videos_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoSubmitRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/videos/id/{video_id_path}/status:
    get:
      tags:
      - Videos
      summary: Processing status
      description: 'Return processing status for the given video.


        Accessible to the video owner (creator) or any moderator.'
      operationId: get_video_status_api_v1_videos_id__video_id_path__status_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: video_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Video Id Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/videos/id/{video_id_path}:
    get:
      tags:
      - Videos
      summary: Video details
      description: Public endpoint returning full video metadata.
      operationId: get_video_details_api_v1_videos_id__video_id_path__get
      parameters:
      - name: video_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Video Id Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Videos
      summary: Update video details
      description: Allow owner or moderator to update title/desc/tags.
      operationId: update_video_api_v1_videos_id__video_id_path__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: video_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Video Id Path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/videos/id/{video_id_path}/view:
    post:
      tags:
      - Videos
      summary: Record playback view
      description: "Increment a video's view count.\n\nBehaviour nuances (aligned\
        \ with test-suite expectations):\n\n\u2022 If the video does **not exist**\
        \ \u2192 404 for everyone.\n\u2022 If the video exists but is **not READY**:\n\
        \    \u2013 **Unauthenticated** callers receive 404 (video hidden).\n    \u2013\
        \ Authenticated *viewer*-level callers receive 403.\n    \u2013 The owner\
        \ (*creator*) or any *moderator* can still access \u2192 404 to\n      remain\
        \ consistent with current spec (not explicitly tested yet).\n\u2022 A READY\
        \ video is public: anyone can record a view (204)."
      operationId: record_view_api_v1_videos_id__video_id_path__view_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: video_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Video Id Path
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/videos/latest:
    get:
      tags:
      - Videos
      summary: Latest videos
      operationId: get_latest_videos_api_v1_videos_latest_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page number
          default: 1
          title: Page
        description: Page number
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Items per page
          default: 10
          title: Pagesize
        description: Items per page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/videos/by-tag/{tag_name}:
    get:
      tags:
      - Videos
      summary: Videos by tag
      operationId: get_videos_by_tag_api_v1_videos_by_tag__tag_name__get
      parameters:
      - name: tag_name
        in: path
        required: true
        schema:
          type: string
          title: Tag Name
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page number
          default: 1
          title: Page
        description: Page number
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Items per page
          default: 10
          title: Pagesize
        description: Items per page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/videos/by-uploader/{uploader_id_path}:
    get:
      tags:
      - Videos
      summary: Videos by uploader
      operationId: get_videos_by_uploader_api_v1_videos_by_uploader__uploader_id_path__get
      parameters:
      - name: uploader_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Uploader Id Path
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page number
          default: 1
          title: Page
        description: Page number
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Items per page
          default: 10
          title: Pagesize
        description: Items per page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/videos/id/{video_id_path}/rating:
    post:
      tags:
      - Videos
      summary: Submit rating (1-5)
      operationId: submit_rating_api_v1_videos_id__video_id_path__rating_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: video_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Video Id Path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoRatingRequest'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Videos
      summary: Get rating summary
      operationId: get_rating_summary_endpoint_api_v1_videos_id__video_id_path__rating_get
      parameters:
      - name: video_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Video Id Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoRatingSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/videos/id/{video_id_path}/related:
    get:
      tags:
      - Videos
      summary: Content-based related list
      description: 'Return a list of videos related to the given video.


        The underlying implementation is currently stubbed out and will return the

        latest videos (excluding the source video) with a random relevance score.'
      operationId: get_related_videos_for_video_api_v1_videos_id__video_id_path__related_get
      parameters:
      - name: video_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Video Id Path
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 20
          minimum: 1
          description: Max number of related videos
          default: 5
          title: Limit
        description: Max number of related videos
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecommendationItem'
                title: Response Get Related Videos For Video Api V1 Videos Id  Video
                  Id Path  Related Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/videos/trending:
    get:
      tags:
      - Videos
      summary: Trending videos (top by views)
      description: "Return the *trending* list \u2013 most viewed videos in the selected\
        \ window."
      operationId: get_trending_videos_api_v1_videos_trending_get
      parameters:
      - name: intervalDays
        in: query
        required: false
        schema:
          type: integer
          maximum: 30
          minimum: 1
          description: Time window in days to consider (1, 7, or 30)
          examples:
            one:
              summary: 1 day
              value: 1
          default: 1
          title: Intervaldays
        description: Time window in days to consider (1, 7, or 30)
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 10
          minimum: 1
          description: Maximum number of records to return (max 10)
          default: 10
          title: Limit
        description: Maximum number of records to return (max 10)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VideoSummary'
                title: Response Get Trending Videos Api V1 Videos Trending Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/videos/preview:
    post:
      tags:
      - Videos
      summary: Preview YouTube URL (title only)
      description: "Return the title of the supplied YouTube video.\n\nNo authentication\
        \ required \u2013 the client uses this to pre-fill the *Name*\nfield when\
        \ submitting a new video.  All heavy-lifting stays backend-side so\nwe avoid\
        \ CORS issues and leaking any API keys."
      operationId: preview_youtube_video_api_v1_videos_preview_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoSubmitRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoPreviewResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/search/videos:
    get:
      tags:
      - Search
      summary: Keyword video search
      operationId: search_videos_api_v1_search_videos_get
      parameters:
      - name: query
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          description: Search query term
          title: Query
        description: Search query term
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page number
          default: 1
          title: Page
        description: Page number
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Items per page
          default: 10
          title: Pagesize
        description: Items per page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    PaginatedResponse:
      properties:
        data:
          items: {}
          type: array
          title: Data
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - data
      - pagination
      title: PaginatedResponse
...
    VideoDetailResponse:
      properties:
        title:
          type: string
          maxLength: 100
          minLength: 3
          title: Title
        description:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Description
        tags:
          items:
            type: string
          type: array
          title: Tags
        videoId:
          type: string
          format: uuid
          title: Videoid
        userId:
          type: string
          format: uuid
          title: Userid
        submittedAt:
          type: string
          format: date-time
          title: Submittedat
        thumbnailUrl:
          anyOf:
          - type: string
            maxLength: 2083
            minLength: 1
            format: uri
          - type: 'null'
          title: Thumbnailurl
        location:
          type: string
          title: Location
        location_type:
          type: integer
          title: Location Type
        content_features:
          anyOf:
          - items:
              type: number
            type: array
          - type: 'null'
          title: Content Features
        content_rating:
          anyOf:
          - type: string
          - type: 'null'
          title: Content Rating
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
        youtubeVideoId:
          anyOf:
          - type: string
          - type: 'null'
          title: Youtubevideoid
        updatedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updatedat
        status:
          $ref: '#/components/schemas/VideoStatusEnum'
          default: PENDING
        views:
          type: integer
          title: Views
          default: 0
        averageRating:
          anyOf:
          - type: number
          - type: 'null'
          title: Averagerating
        totalRatingsCount:
          type: integer
          title: Totalratingscount
          default: 0
        is_deleted:
          type: boolean
          title: Is Deleted
          default: false
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - title
      - userId
      - submittedAt
      - location
      - location_type
      title: VideoDetailResponse
      description: Response model returned when fetching full video details.
    VideoPreviewResponse:
      properties:
        title:
          type: string
          title: Title
      type: object
      required:
      - title
      title: VideoPreviewResponse
      description: Response model used by the *preview* endpoint to pre-fill UI fields.
    VideoRatingRequest:
      properties:
        rating:
          type: integer
          maximum: 5.0
          minimum: 1.0
          title: Rating
      type: object
      required:
      - rating
      title: VideoRatingRequest
      description: Client payload for submitting a rating (1-5).
    VideoRatingSummary:
      properties:
        videoId:
          type: string
          format: uuid
          title: Videoid
        averageRating:
          type: number
          title: Averagerating
        ratingCount:
          type: integer
          title: Ratingcount
      type: object
      required:
      - videoId
      - averageRating
      - ratingCount
      title: VideoRatingSummary
      description: Aggregated rating stats for a video.
    VideoStatusEnum:
      type: string
      enum:
      - PENDING
      - PROCESSING
      - READY
      - ERROR
      title: VideoStatusEnum
      description: Possible processing states for a submitted video.
    VideoStatusResponse:
      properties:
        videoId:
          type: string
          format: uuid
          title: Videoid
        status:
          $ref: '#/components/schemas/VideoStatusEnum'
      type: object
      required:
      - videoId
      - status
      title: VideoStatusResponse
      description: Response model that surfaces only the processing status.
    VideoSubmitRequest:
      properties:
        youtubeUrl:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Youtubeurl
        title:
          anyOf:
          - type: string
            maxLength: 150
            minLength: 1
          - type: 'null'
          title: Title
          description: Optional video title obtained from the preview step.
      type: object
      required:
      - youtubeUrl
      title: VideoSubmitRequest
      description: "Payload accepted by the *submit video* endpoint.\n\nThe **title**\
        \ is optional \u2013 if the client already looked it up via the\npreview endpoint\
        \ it can pass it here so the user immediately sees the\ncorrect name instead\
        \ of the temporary placeholder."
...
    VideoUpdateRequest:
      properties:
        title:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 3
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Description
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tags
      type: object
      title: VideoUpdateRequest
      description: Payload for partial updates to a video owned by the caller or a
        moderator.
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/users/login